main.rbからothello.rbを分割
